+++ zentrale/wetter.h
@@ -19,7 +19,7 @@ namespace Local {
 		}
 
 		int gib_stundenvorhersage_solarstrahlung_in_prozent(int index, Local::Config& cfg) {
-			return round(stundenvorhersage_solarstrahlung_liste[index] * 100 / cfg.maximale_solarstrahlung_in_w_pro_m2);
+			return round((float) stundenvorhersage_solarstrahlung_liste[index] * 100 / (float) cfg.maximale_solarstrahlung_in_w_pro_m2);
 		}
 
 		void setze_tagesvorhersage_solarstrahlung(int index, int val) {
@@ -27,7 +27,7 @@ namespace Local {
 		}
 
 		int gib_tagesvorhersage_solarstrahlung_in_prozent(int index, Local::Config& cfg) {
-			return round(tagesvorhersage_solarstrahlung_liste[index] * 100 / cfg.maximale_solarstrahlung_pro_tag_in_w_pro_m2);
+			return round((float) tagesvorhersage_solarstrahlung_liste[index] * 100 / (float) cfg.maximale_solarstrahlung_pro_tag_in_w_pro_m2);
 		}
 
 		void set_log_data(char* buffer) {
